chore(new-issuer): plug presenter code to containers +
refactor dividend card as separate component
