+++ packages/new-polymath-issuer/src/components/Checkpoints/styles.ts
@@ -4,7 +4,6 @@ import {
   List,
   Heading,
   IconOutlined,
-  Button,
 } from '@polymathnetwork/new-ui';
 
 export const iconSize = 25;
@@ -53,9 +52,4 @@ export const StepIcon = styled(IconOutlined)`
   top: -${iconSize}px;
   left: 0;
   margin-left: -${iconSize / 2}px;
-`;
-
-export const NewDividendButton = styled(Button)`
-  flex-direction: column;
-  align-items: center;
 `;
