+++ packages/new-polymath-issuer/src/components/DividendCard/index.ts
@@ -1,0 +1,1 @@
+export * from './DividendCard';
